+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unrelease]
+
+### Chnaged
+- UI improvements on amount selection
+
 ## [4.0.11] - 2023-08-11
 
 ### Changed